refactor: avoid chrome api running in nodejs
