+++ packages/extension-chrome/src/services/storage/index.ts
@@ -1,0 +1,1 @@
+export { createInMemoryStorage } from './memory';
