+++ packages/extension-chrome/src/services/storage/memory.ts
@@ -1,37 +1,10 @@
 import { Storage } from '@nexus-wallet/types';
 import { errors } from '@nexus-wallet/utils';
-import browser from 'webextension-polyfill';
 
 interface InMemoryStorage<S> extends Storage<S> {
   getAll(): S | undefined;
 
   setAll(s: S): void;
-}
-
-export function createBrowserExtensionStorage<S>(): Storage<S> {
-  return {
-    getItem: (key) => {
-      const k = String(key);
-      return browser.storage.local.get(k).then((record) => record[k]);
-    },
-    removeItem: async (key) => {
-      const k = String(key);
-      const obj = await browser.storage.local.get(k);
-      if (k in obj) {
-        await browser.storage.local.remove(k);
-        return true;
-      }
-      return false;
-    },
-    setItem: (key, value) => {
-      return browser.storage.local.set({ [key]: value });
-    },
-    hasItem: async (key) => {
-      const k = String(key);
-      const obj = await browser.storage.local.get(k);
-      return k in obj;
-    },
-  };
 }
 
 export function createInMemoryStorage<S>(): InMemoryStorage<S> {
